Refactor createParentMixin
