+++ tsconfig.json
@@ -37,12 +37,14 @@
 		"./src/mixins/createFormFieldMixin.ts",
 		"./src/mixins/createListMixin.ts",
 		"./src/mixins/createObservedListMixin.ts",
+		"./src/mixins/createParentMapMixin.ts",
 		"./src/mixins/createParentMixin.ts",
 		"./src/mixins/createRenderable.ts",
 		"./src/mixins/createRenderableChildrenMixin.ts",
 		"./src/mixins/createStatefulChildrenMixin.ts",
 		"./src/mixins/createTabbedMixin.ts",
 		"./src/mixins/createVNodeEvented.ts",
+		"./src/mixins/interfaces.d.ts",
 		"./src/projector.ts",
 		"./src/util/createCancelableEvent.ts",
 		"./src/util/createMemoryStore.ts",
@@ -75,6 +77,7 @@
 		"./tests/unit/mixins/createCloseableMixin.ts",
 		"./tests/unit/mixins/createFormFieldMixin.ts",
 		"./tests/unit/mixins/createListMixin.ts",
+		"./tests/unit/mixins/createParentMapMixin.ts",
 		"./tests/unit/mixins/createParentMixin.ts",
 		"./tests/unit/mixins/createRenderable.ts",
 		"./tests/unit/mixins/createStatefulChildrenMixin.ts",
